docker_top/requirements.txt
python-telegram-bot>=20.0
requests
dotenv
